refactor(vsock-proto): split command payload module
