+++ crates/vsock-proto/src/payloads/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod command;
